# specs/001-fullstack-todo/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API
  description: RESTful API for Phase II Fullstack Todo Application
  version: 1.0.0
  contact:
    name: Desktop Todo Project

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /api/todos:
    get:
      summary: List all todos
      description: Retrieve all todos ordered by creation date (newest first)
      operationId: listTodos
      tags:
        - Todos
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoRead'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new todo
      description: Create a new todo with title and optional description
      operationId: createTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Todo UUID
        schema:
          type: string
          format: uuid

    get:
      summary: Get a specific todo
      description: Retrieve a single todo by its ID
      operationId: getTodo
      tags:
        - Todos
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a todo
      description: Update title and/or description of an existing todo
      operationId: updateTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a todo
      description: Permanently delete a todo by its ID
      operationId: deleteTodo
      tags:
        - Todos
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/todos/{id}/toggle:
    parameters:
      - name: id
        in: path
        required: true
        description: Todo UUID
        schema:
          type: string
          format: uuid

    patch:
      summary: Toggle todo completion status
      description: Toggle the completed status of a todo (true → false or false → true)
      operationId: toggleTodo
      tags:
        - Todos
      responses:
        '200':
          description: Todo toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    TodoCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          description: Task title (required)
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional detailed description
          example: Milk, eggs, bread, butter

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          description: Updated task title
          example: Buy groceries and snacks
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Updated description
          example: Milk, eggs, bread, butter, chips

    TodoRead:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          maxLength: 255
          description: Task title
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional detailed description
          example: Milk, eggs, bread, butter
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: '2026-01-21T10:30:00Z'

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Todo not found

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Location of the error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type

tags:
  - name: Todos
    description: Todo CRUD operations